Edited name structures
