+++ problem_2/ea/__init__.py
@@ -1,0 +1,3 @@
+from ea.strategy import Strategy
+from problem_2.ea.ea_factory import EvolutionaryFactory
+from ea.individual import Individual
