--- problem_2/es/__init__.py
@@ -1,3 +1,0 @@
-from es.strategy import Strategy
-from es.es_factory import EvolutionaryFactory
-from es.individual import Individual
