+++ problem_2/evaluators/abstraction/fitness_evaluator.py
@@ -1,5 +1,5 @@
 from abc import ABC, abstractmethod
-from es.individual import Individual
+from ea.individual import Individual
 
 class FitnessEvaluator(ABC):
     @abstractmethod
