+++ problem_2/recombinators/abstraction/recombinator.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 from typing import List
-from es.individual import Individual
+from ea.individual import Individual
 
 class Recombinator(ABC):
     @abstractmethod
